improves convention related tests

+++ arbiter/src/main/java/com/roquentin/arbiter/services/ConventionService.java
@@ -8,7 +8,6 @@ import org.springframework.stereotype.Service;
 import com.roquentin.arbiter.dto.ConventionDTO;
 import com.roquentin.arbiter.expections.UnauthorizedException;
 import com.roquentin.arbiter.models.Convention;
-import com.roquentin.arbiter.models.Cooperation;
 import com.roquentin.arbiter.repositories.ConventionRepository;
 
 @Service
@@ -24,17 +23,21 @@ public class ConventionService {
 	private UserService userService;
 	
 	@Transactional
-	public Convention createConvention(ConventionDTO newConvention) {
+	public boolean createConvention(ConventionDTO newConvention) {
 		//TODO:: voting
 		if (!cooperationService.canUserChangeCooperation(newConvention.getCooperationId(), userService.getCurrentUser()))
 			throw new UnauthorizedException("No permissions to change the cooperation");
 		
-		return repository.saveUsingDTO(newConvention.getName(), newConvention.getDescription(),
+		return 0 < repository.saveUsingDTO(newConvention.getName(), newConvention.getDescription(),
 				newConvention.getCooperationId(), newConvention.getConsequence());
 		
 	}
 	
 	public void deleteConvention(Long id) {
+		Convention convention = repository.getOne(id);
+		if (!cooperationService.canUserChangeCooperation(convention.getCooperation(), userService.getCurrentUser()))
+			throw new UnauthorizedException("No permissions to change the cooperation");
+		
 		//TODO: voting
 		repository.deleteById(id);
 	}
